+++ dolfin_adjoint_common/blocks/__init__.py
@@ -2,8 +2,8 @@ from .solving import SolveBlock
 from .assembly import AssembleBlock
 from .projection import ProjectBlock
 from .solving import SolveBlock
-from variational_solver import NonlinearVariationalSolveBlock
-from .function import FunctionEvalBlock, FunctionAssignBlock, FunctionSplitBlock, FunctionMergeBlock
+from .variational_solver import NonlinearVariationalSolveBlock
+from .function import FunctionAssignBlock
 from .function_assigner import FunctionAssignerBlock
 from .mesh import ALEMoveBlock, BoundaryMeshBlock
 from .expression import ExpressionBlock
