+++ dolfin_adjoint_common/blocks/constant.py
@@ -1,3 +1,5 @@
+from pyadjoint import Block
+
 
 class ConstantAssignBlock(Block):
     def __init__(self, func, other):
