+++ dolfin_adjoint_common/blocks/dirichlet_bc.py
@@ -1,3 +1,5 @@
+from pyadjoint import Block, no_annotations
+
 
 class DirichletBCBlock(Block):
     def __init__(self, *args):
