+++ dolfin_adjoint_common/blocks/expression.py
@@ -1,3 +1,5 @@
+from pyadjoint import Block
+
 
 class ExpressionBlock(Block):
     def __init__(self, expression):
