+++ dolfin_adjoint_common/blocks/function_assigner.py
@@ -1,3 +1,5 @@
+from pyadjoint import Block
+
 
 class FunctionAssignerBlock(Block):
     def __init__(self, assigner, inputs):
