+++ fenics_adjoint/blocks/__init__.py
@@ -1,5 +1,6 @@
 from .common import *
 from .variational_solver import LinearVariationalSolveBlock
 from .krylov_solver import KrylovSolveBlock
-from .lu_solver import LUSolveBlock
-from .petsc_krylov_solver import PETScKrylovSolveBlock
+from .lu_solver import LUSolveBlock, LUSolveBlockHelper
+from .petsc_krylov_solver import PETScKrylovSolveBlock, PETScKrylovSolveBlockHelper
+from fenics_adjoint.blocks.function import FunctionEvalBlock, FunctionSplitBlock, FunctionMergeBlock
