+++ fenics_adjoint/types/dirichletbc.py
@@ -14,7 +14,7 @@ from pyadjoint.block import Block
 #       new boundary values/function.
 
 
-class DirichletBC(FloatingType, compat.DirichletBC):
+class DirichletBC(FloatingType, backend.DirichletBC):
     def __init__(self, *args, **kwargs):
         super(DirichletBC, self).__init__(*args, **kwargs)
