+++ fenics_adjoint/types/function_space.py
@@ -1,5 +1,8 @@
+import backend
 from dolfin_adjoint_common import compat
 compat = compat.compat(backend)
+
+
 extract_subfunction = compat.extract_subfunction
 
 __all__ = ["extract_subfunction"]
